初步测试game系统正常

+++ GameGraphics/SceneCommon.cpp
@@ -6,7 +6,7 @@ SceneCommon::SceneCommon() {}
 
 SceneCommon::~SceneCommon() {}
 
-void SceneCommon::update(long long dt) {}
+void SceneCommon::update(float dt) {}
 
 bool SceneCommon::handleEvent(const sf::Event& evt) {
     return false;
